Blanket scale down all images in LaTeX build (fix #55)
Most of the screenshots were too wide for the format (450px ~= 451 pts, when
the colums are 417 pts wide), so a simple inplace sed operation was added
to scale down every image by 8%, reducing the overfull hboxes by about as
much as can be expected.

Though there are still places where it overflows, the remaining places are
mostly sub-point overshoots or in a few cases, standalone lines with <5pts
overshoot. I.e. hardly noticable.